test: add tests for new components and some refactoring
