+++ src/stores/calendarStore.test.ts
@@ -4,7 +4,7 @@
 
 import dayjs from "dayjs";
 import useCustomFormat from "dayjs/plugin/customParseFormat";
-import mockCalendar from "../../__mocks__/calendar.json";
+import mockCalendar from "../__mocks__/calendar.json";
 import CalendarStore from "./calendarStore";
 
 dayjs.extend(useCustomFormat);
